ENH: Add benchmarking with airspeed velocity
Benchmarks for creating databases from a string and from a file object, creating
a model from a database, generating Halton sequences of varying dimension and
number of points, running calculate on an example set, and running equilibrium
on two example sets are added.

These benchmarks are not comprehensive, but serve the purpose of showing changes
in pycalphad's general performance over time.

The benchmarks were split up so the equilibrium benchmarks have their own class.
The reason for this is so that they can run on the wallclock time and so that
they can be run a specific number of times (5) (because they are long, they would
only be run once by default).

Changes per comments in gh-48 for asv integration

README.rst: Fix typo in asv link

RELEASING.rst: Add a step with instructions for running the benchmarks
for a new tagged release. Publishing to gh-pages will be done using CI
(see gh-49).

benchmarks.py: Add magnetic and non-magnetic model construction and
calculate() commands. The FeC tdb and database were removed because they
were unused after changing calculate() to use AlFe. Dead code that
created models in the setup was removed.

asv.conf.json: Change target branch to develop from develop and master
because the master should be done by hand during the release procedure
and fine-grain benchmarks should be done in the develop branch.