Add Binder link to README.rst
I was able to go into the examples and run the BinaryExamples in Binder, but it was not picking up and installing pycalphad correctly. Running the following command in the Jupyter Notebook to install it worked fine.
`! conda install -y -p /home/main/anaconda2/envs/python3 -c conda-forge pycalphad`

The performance isn't bad either. I was able to run the Al-Zn example in about 2 minutes on my i7 MacBook and it took about 5 to run in the Docker instance.