BLD/TST: Add support for Python 3.6
Closes gh-66.

Add Python 3.6 to Travis and Appveyor

Python 3.6 changed `splitext` to support path-like objects, so passing a StringIO
(file-like) now throws a `TypeError` instead of an `AttributeError`. This was
changed at 3 instances in the format determination.

A test was also written to check that passing a StringIO directly to the Database
constructor (e.g. not using the static method constructors) will proceed
correctly through the try-except block. It's the same test as from_file, it just
did not exist for the direct __new__ constructor.