FIX: Create a blacklist for compiler files
Compiler settings on certain machines can create files such as debug
symbols. This commit creates a blacklist for filtering those files out,
leaving only the compiled file. We cannot more directly find the compiled
file because only the extensions differ (e.g. '*.so' and '*.so.dysm') and
the extensions of the compiled files change on different platforms.

Fixes gh-65