Update license
This is a license year bump and also adds @bocklund as an identified contributor to the project. (My employer, Caltech, is also added to the copyright notice, per a software release agreement.)
It's probably a good idea to get @bocklund on record that, while he retains the copyright over his own contributions (no CLA), he is aware that all of his contributions are made to the project under the terms of the MIT license. Do you consent to this?