ENH: TDB compatibility: all chars after command delimiters should be neglected
Thermo-Calc will not parse commands or other strings after the command
delimiter. Previously pycalphad only removed comments after command delimiters.
This commit makes pycalphad conform to the behavior of parsing out everything
on the line after the delimiter and includes a test for this behavior.

Another approach might be to only strip characters after the *last* delimiter
on a line, which would be an overall improved approach in the case that
multiple commands were on a single line (or an entire TDB on a single line)
even though I have never encountered such a case in the wild. I'm not sure how
other commercial software handles multiple commands per line.