ENH: Species support in Database and TDB read/write (#137)
Species in a TDB are parsed in the `Specie` class and added to the `species` set attribute of the Database.

The Specie class has several attributes:

- `name`: the name of the specie
- `constituents`: a dictionary of `{element: composition}` of the specie
- `charge`: integer charge. Defaults to zero if none specified.

As discussed in the Thermo-Calc database manager guide, elements entered via the ELEMENT command are automatically entered as species. We match that behavior here. The constituents of an element are just that element with a fraction of 1 and the charge is 0.

In principle, the constituents dictionary will be used in the solver to link the specie to how it contributes to the overall mass balance conditions.

Tests are included that test the basic functionality.