FIX: fix Species Python 2 unicode support (#166)
Python2 can have strings with different types, including `unicode`. Unicode strings do not match the `isinstance(obj, str)`, so here a workaround was implemented in the spirit of the `six` package that checks the version of Python and either uses `basestring` or `str` for the string type for Py2 and Py3, respectively.

A test was included.