FIX: Max phases by Gibbs phase rule accommodated  (#184)
* Mostly straightforward fix that closes #170 by padding the `vertex` to be of size `C+2` instead of `C` (for C components). 
* Includes a test of a eutectic point in Pb-Sn. 
* Modestly improved hyperplane documentation

This requires a hack in `hyperplane` because hyperplane constructs a C-simplex hyperplane corresponding to the chemical potentials, however the phase fraction and point index arrays are now padded to accommodate C+2 phases instead of C phases. So I set the point index to zero (which pulls in the phase, site fracs, etc. associated with that point) and the phase fraction of those points to zero.