ENH: Reading large databases via delayed parameter processing (#266)
Database('steel_database_fix.tdb') can take a long time to complete, 40-90 seconds depending on the machine. Switching to tinydb's insert_multiple API is a cheap and easy way to cut that time in ~half.

Most of the remaining time is in parsing the grammar for all the PARAMETERs, as well as calls to sympify for the function expressions. There's room to improve there (may be possible to get the time down to ~6 seconds), but those changes may break some internal or external APIs and will be more complicated to implement. This is the low-hanging fruit

* WIP/ENH: tdb/Database: Performance tuning for large databases

* FIX: tdb: Accidentally removed grammar

* MNT: tdb/Database: Rename Database.process_parameter_queue() and make it a public API