Fix writing Piecewise intervals for cases where TDBs have an interval size of zero (see text):
```
FUNCTION UN_ASS 300 0; 300 N !
```

Will fail because the start and end intervals are the same and SymPy will not make this a Piecewise function.

This could be troublesome for a function

```
FUNCTION TENK 300 10000; 300 N !
```

which is non-zero. The commercial codes will extrapolate above/below, but SymPy will simplifiy this to an empty Piecewise() (where pycalphad appends a (0, True) expression).

```python
>>> Piecewise((10000, And(v.T >= 300, v.T < 300)))
Piecewise()
```