MAINT: Rename models -> model for backwards compatibility (#272)
In reviewing #270, I overlooked the fact that `calculate` takes the argument `model`, which a caller could pass as a keyword argument to `_eqcalculate`. #270 changed `_eqcalculate` to explicitly take a `models` argument and pass it to `calculate`.

This preserves the behavior from #270 and stays backwards compatible if anyone was using `model` as a keyword argument that got passed through to `calculate`.