FIX: Model: improve _array_validity (#258)
Fixes a bug where Database parameters would be consumed by `Model._array_validity` without requiring that the species is in the Phase constituents (#256). This closes #256 and includes a test to prevent any future regressions.

There's also some refactoring of `Model._purity_test` and `Model._interaction_test` to rely on `Model._array_validity` to determine whether the constituent array is valid, and these functions can be left to simply test for whether or not there are any interactions.