MAINT: Fix NumPy deprecation warnings (#312)
NumPy 1.20 deprecated the use of `np.int`, `np.float`, `np.complex`, ... ([see here](https://numpy.org/devdocs/release/1.20.0-notes.html#using-the-aliases-of-builtin-types-like-np-int-is-deprecated)) which are type aliases that shadow the built-in types `int`, `float`, `complex`, ...

This PR was the result of 3 regex replaces on `.py`, `.pyx`, and `.pxd` files:

1. `\bnp.int\b` → `np.int_`
2. `\bnp.float\b` → `np.float_`
3. `\bnp.complex\b` → `np.complex_`

These aliases have been around a long time (they are present in our minimum NumPy version 1.13 - tested on NumPy 1.13.3) so this change is backwards compatible.  The trailing `_` versions no longer shadow built-in types and keeping the suffixed versions allows us to find/replace these in the future (compared to built-in types, which we cannot).

For future reference:  `pytest -v pycalphad -W error::DeprecationWarning` can be used to raise errors on DeprecationWarnings. Doing this resulted in 70 tests failing before this PR. Now no tests fail.