FIX: add more TDB reference key characters (#326)
This PR fixes a regression introduced in #308 where the reference key in TDB `FUNCTION` and `PARAMETER` lines was limited to alphanumeric characters. Now `:_-` characters as well to fix some local databases that fail to parse. It's unclear what other valid characters could be. Thermo-Calc gives examples in the documentation that include `REF:250`, `REF_002`, and `REF-SGTE`, so at least those are confirmed valid characters.

On my on the 522 local TDB files, this dropped the parse failures from 41 to 14. The 14 failures to parse are due to whitespace in the reference key, e.g. `PARAM ...; 6000 N REF: 0 !`, which is invalid per Thermo-Calc's database checker.