take_step: always correct the site fractions, statevars and phase amounts
This is a slight code simplification. It also may leave the system in
an inconsistent state. The chemical potentials may not be consistent
with the phase amounts, state variables and site fractions, since they
were corrected from the result of the equilibrium solution. In the limit
of convergence, it might not matter, since all the corrections should
be converged below our numerical tolerance anyway.