DOC: Fix docs highlighting issue stemming from incompatible packages (#353)
The CI environment had an old version of pip that used the old resolver that would let incompatible packages be installed. In this case, furo was incompatible with sphinx 4, but sphinx 4 was installed anyway, which led to breakage in the code highlighting. Upgrading pip first fixes the issue because the newer pip resolver disallows installing incompatible versions.

This PR also ensures that we use `pip install -U` everywhere to get the latest versions of everything.