Pull species out of the solver
This implies a tighter, explicit coupling between the solver
and the CompositionSet/PhaseRecords objects. We make the assumption that
PhaseRecord.nonvacant_elements are the non-vacanct pure elements for the
_system_, not just the local phase it describes. This was already
implictly happening in build_phase_records and it's unclear whether
there was implicit tight coupling.

The main thing here is that this _explictly_ disallows mixing
PhaseRecord objects built with different subsets of the system marked
as "active".