Remove some TODOs for sorting order of RKM parameters after a test
The experiment was to do calculations in Thermochimica for a simple database with a phase that had constant energy excetp for an L1 term.
There were for tested cases for the test matrix of:
- Endmembers ("Pb", "Sn") and endmbmers ("Sn", "Pb")
- Interacting species indices (1, 2) and (2, 1)

The results of the experiment were that switching the order of endmembers or the order of interacting species indices caused the effect of the L1 term to change sign (changing both was the same as changing neither, -1 * -1 = 1). The ordering in the database _is_ important for odd order excess parameters.