FIX: Make minimizer state more internally consistent (#410)
This PR moves `advance_state` to be the last step in the minimization loop. 

The main reason to make this change is that we want the chemical potentials that we get from `solve_state` to be consistent with the current site fractions, phase amounts, and state variables when we check for convergence, change phases, and return a solution. Right now, we apply the corrections to site fractions, phase amounts, and state variables after solving for the chemical potentials with the current values, so they are effectively one step ahead of the chemical potentials. 

This fixes #372 as the driving forces should now always be zero in `change_phases`, as the chemical potentials are now consistent with the current site fractions. 

Summary of changes:
1. If the phases change, we don't advance the state. It probably makes sense to be doing this so that we only change one thing at a time (i.e. we don't want to change the site fractions/phase amounts/state variables _and_ the phases).
2. We set the numerical limit to phase amounts in the adaptive step size determination in `advance_state` to something non-zero to avoid divide by zero cases.
3. Slightly adjust the convergence tolerance. Without this adjustment, the rose test was failing. Because we were returning site fractions for the step ahead, I think the root cause was that the next step would converge to the correct energy in the test, so slightly tightening the tolerances seemed to help get there. Still, the limiting factor for something like `ALLOWED_DELTA_Y = 1e-09` seems to be the charged species test.