ENH: Add pathlib support to Database reading (#545)
Enables the following behavior:

```python
from pathlib import Path
from pycalphad import Workspace
wks = Workspace()
wks.database = Path("dat/Fe-C.dat")
```

Closes #544 

We test this by changing the tests that usually wrap `files()` output (a `PosixPath`) in a `str()` to make a string-like path. We remove this behavior in several tests.

This also changes some tests (`test_unspecified_format_from_string` and `test_unspecified_format_from_file`) to be more in line with what I interpreted to be the spirit of the tests: passing database strings (not strings for the paths to the databases) to `Database.from_string` or `Database.from_file` should fail if a format is not specified.