FIX: Better handle convergence failures in Workspace (#558)
This PR adds a testing `ConvergenceFailureSolver` to the test suite that overrides `solve` to return an unconverged result. We use this solver to add test cases for situations where there are convergence failures without having to track and continually update the test suite with real convergence failures as the solver improves to fix them.

In general, the expected behavior for convergence failures is that all properties derived from equilibrium return `NaN`, which is backwards compatible with pre-Workspace xarray Datasets. 

In particular, we test and fix bugs where the following `wks.get` calls would raise errors by trying to access composition sets that don't exist, taking into account the changes in #540 for `find_first_compset`:
- `wks.get("MU(component)")`
- `wks.get("Y(phase_name, sublattice_index, species)")`
- `wks.get(IsolatedPhase("phasename", wks=wks)("property"))`
- `wks.get(DormantPhase("phasename", wks=wks)("property"))`