FIX: Fix lambdifying models that evaluate to have complex numbers (#599)
In PyCalphad we eventually represent all our exponents as floating points, even if they are actually integer values. This becomes troublesome because SymEngine evaluates small, negligible imaginary parts. When that happens, the imaginary part is not useful and causes an issue when we call `lambdify` in (`build_functions`) with the default argument that `real=True`. We probably haven't run into this before because our bases are typically symbolic (e.g. `T` in power functions) and these are typically not evaluated to a complex number until the function is actually evaluated inside `lambdify` (where the imaginary part gets thrown away because we call `lambdify(..., real=True)` implicitly). Consequently, doing this replacement outside of `lambdify` can't be a source of a regression because `lambdify` has been throwing away any complex parts anyway. 

We end up doing this workaround to find complex numbers by `atoms` and replace the imaginary unit with zero, which is working around SymEngine not having a `re` (real) function and calling `subs` or `xreplace` with `{I: 0.0}` directly on the Model expression doesn't replace arbitrarily nested imaginary numbers.