FIX: Fix `_compute_phase_values` adding a point with garbage data when the size of feasible `points` is zero. (#627)
This PR fixes the bug described in #626 where `_compute_phase_values` produces a point with bad data when there are no `points` from `_sample_phase_constitution` (i.e. `points.size == 0`). 

The cause of the bug is that `np.atleast_2d(points)` gives `points.shape == (1, 0)` in that case, which messes up the logic in the code that assumes `points.shape[0]` refers to the number of points, which it usually does except for when the `points` array is size zero and the sizes of the dimensions are meaningless. This causes calls to `phase_record.prop_2d`/`prop_parameters_2d`  and `phase_record.mass_obj_2d` to produce garbage data because the `dof` provided is malformed.

The fix implemented here is to only call into the phase record functions if there are a non-zero number of points and otherwise make empty, zero-sized `phase_output` and `phase_compositions` arrays that have the correct shapes (and to correct the shape of `points` as it is used later).

If a call to `calculate` produces no feasible points, each phase that produces no points will raise a warning and if there are no feasible points among all phases a `ConditionError` is raised that prompts the user to check the active phases and set of conditions (and state variables).